large-language-models/research-assistants/text-to-sql/requirements.txt
langchain-community==0.3.17
langchain-openai==0.3.5
langgraph==0.2.72
pandas==2.2.3
